# server-api-generator/src/main/resources/openapi.yaml
openapi: 3.0.2
...
info:
  title: jrc-node server api spec
  description: "OpenAPI specification for the blockchain server backend for jrc-node"
  license:
    name: MIT
    url: https://opensource.org/licenses/mit-license.php
  version: 0.0.1-SNAPSHOT
...
paths:
...
  /blockchain/addblock:
    post:
      summary: Attempt to add a new block to the blockchain
      operationId: addBlock
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
        400:
          description: block is not valid
          content:
            application/json:
              schema:
                type: object
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Block'
        description: Block to be added
        required: true
...
components:
  schemas:
...
    Block:
      type: object
      properties:
        hash:
          type: string
        previousHash:
          type: string
        data:
          type: string
        timeStamp:
          type: string
        nonce:
          type: string
        difficulty:
          type: string
        proofOfWork:
          type: string